Use the modtime on getdown.txt as a crude locking mechanism to keep multiple instances of getdown from stepping on each others toes.  Getdown now checks the modtime of getdown.txt when it starts, during its downloading, and before launching the real app.  If at any of these points the modtime has changed from when it was last checked, getdown assumes that means another instance is running and bails to let the other instance do its magic.
Mainly this is for a delayed, silent installer that wants to launch after its
delay, but not if getdown has been run manually after it was launched. Also, if
someone manages to get a little click happy on their installer, this will keep
multiple instances from fighting for the bandwidth to download the same file
simultaneously. 



git-svn-id: http://getdown.googlecode.com/svn/trunk@238 a0402a4a-d63b-ad04-3ffe-6d551d543dc1