Using IOException to communicate a desire to abort the download is a little pesky and pushes annoying requirements on users of the Getdown download services (like Yohoho and ToyBox).
Instead we'll allow Downloader.download() to return false to indicate that the
download was aborted by an observer, and the observers will simply return false
to indicate that they wish the download to be aborted.

While I was in there, I moved the Downloader and its concrete implementations
into a getdown.net package because it annoyed me that they were in
getdown.launcher.


git-svn-id: http://getdown.googlecode.com/svn/trunk@242 a0402a4a-d63b-ad04-3ffe-6d551d543dc1