Reenable locking, but only when -Dlock=true is passed on the command line or a lock="true" param is passed to the applet.
Release the lock when launching the application directly from getdown,
navigating away from the applet and borking.  Hopefully one of the first two
releases will allow invoking directly to work in an applet, but applets seem to
share the lock on my Mac, so I haven't been able to test it.



git-svn-id: http://getdown.googlecode.com/svn/trunk@247 a0402a4a-d63b-ad04-3ffe-6d551d543dc1