This should be with the previous commit which has a lovely blank commit message.
We want to put everything into one jar file which we'll publish via Maven. So
we include our command line tools, and replace the un-proguarded jar with the
proguarded one during the packaging step.

I may look into not renaming the Getdown code so that apps (like Bang) that
link against some of it, can use this jar file. Or I'll just figure out
specifically what apps are likely to link against and not rename just that.
It's probably not a huge size increase either way.


git-svn-id: http://getdown.googlecode.com/svn/trunk@323 a0402a4a-d63b-ad04-3ffe-6d551d543dc1